# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Igloo API",
    "version": "0.1.0",
    "description": "HTTP API for the Igloo self-hosted media center. Most JSON endpoints return a shared envelope with error, optional message, and optional data fields."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin Igloo server"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    }
  ],
...
  "paths": {
...
    "/api/user/password": {
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Update the current user's password",
        "operationId": "updateUserPassword",
        "requestBody": {
          "$ref": "#/components/requestBodies/UpdateUserPasswordRequest"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/MessageSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/api/admin/users/": {
      "get": {
        "tags": [
          "Admin Users"
        ],
        "summary": "List all users",
        "description": "Admin-only endpoint.",
        "operationId": "adminGetUsers",
        "responses": {
          "200": {
            "$ref": "#/components/responses/AdminUsersResponse"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": [
          "Admin Users"
        ],
        "summary": "Create a user",
        "description": "Admin-only endpoint.",
        "operationId": "adminCreateUser",
        "requestBody": {
          "$ref": "#/components/requestBodies/AdminCreateUserRequest"
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/AdminUserResponse"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/admin/users/{id}": {
      "patch": {
        "tags": [
          "Admin Users"
        ],
        "summary": "Update a user",
        "description": "Admin-only endpoint.",
        "operationId": "adminUpdateUser",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPath"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/AdminUpdateUserRequest"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/AdminUserResponse"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin Users"
        ],
        "summary": "Delete a user",
        "description": "Admin-only endpoint.",
        "operationId": "adminDeleteUser",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPath"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/MessageSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/admin/users/{id}/password": {
      "put": {
        "tags": [
          "Admin Users"
        ],
        "summary": "Reset a user's password",
        "description": "Admin-only endpoint.",
        "operationId": "adminResetUserPassword",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPath"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/AdminResetUserPasswordRequest"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/MessageSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Server-side session cookie set by POST /api/auth/login."
      }
    },
    "parameters": {
      "IdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "format": "int64",
          "minimum": 1
        }
      },
...
    },
    "requestBodies": {
...
      "UpdateUserPasswordRequest": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/UpdateUserPasswordRequest"
            }
          }
        }
      },
...
      "AdminCreateUserRequest": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AdminCreateUserRequest"
            }
          }
        }
      },
      "AdminUpdateUserRequest": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AdminUpdateUserRequest"
            }
          }
        }
      },
      "AdminResetUserPasswordRequest": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AdminResetUserPasswordRequest"
            }
          }
        }
      },
...
    },
    "responses": {
...
      "MessageSuccess": {
        "description": "Successful JSON response with an optional message.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MessageSuccess"
            }
          }
        }
      },
...
      "AdminUserResponse": {
        "description": "Admin user response.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AdminUserEnvelope"
            }
          }
        }
      },
      "AdminUsersResponse": {
        "description": "Admin users list response.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AdminUsersEnvelope"
            }
          }
        }
      },
...
      "BadRequest": {
        "description": "Bad request.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid session.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated user is not allowed to perform this action.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Conflict": {
        "description": "Request conflicts with existing data.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Unexpected server error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
      "JsonSuccess": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "boolean",
            "const": false
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "additionalProperties": false
      },
      "MessageSuccess": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "boolean",
            "const": false
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "boolean",
            "const": true
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateUserPasswordRequest": {
        "type": "object",
        "required": [
          "current_password",
          "new_password"
        ],
        "properties": {
          "current_password": {
            "type": "string",
            "format": "password"
          },
          "new_password": {
            "type": "string",
            "format": "password",
            "minLength": 9,
            "maxLength": 128
          }
        },
        "additionalProperties": false
      },
...
      "AuthUser": {
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "is_admin",
          "avatar",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "is_admin": {
            "type": "boolean"
          },
          "avatar": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "additionalProperties": true
      },
...
      "AdminUser": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AuthUser"
          }
        ]
      },
      "AdminCreateUserRequest": {
        "type": "object",
        "required": [
          "name",
          "email",
          "password",
          "is_admin"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "format": "password",
            "minLength": 9,
            "maxLength": 128
          },
          "is_admin": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "AdminUpdateUserRequest": {
        "type": "object",
        "required": [
          "name",
          "email",
          "is_admin"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "is_admin": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "AdminResetUserPasswordRequest": {
        "type": "object",
        "required": [
          "password"
        ],
        "properties": {
          "password": {
            "type": "string",
            "format": "password",
            "minLength": 9,
            "maxLength": 128
          }
        },
        "additionalProperties": false
      },
      "AdminUserEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/JsonSuccess"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "user"
                ],
                "properties": {
                  "user": {
                    "$ref": "#/components/schemas/AdminUser"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        ]
      },
      "AdminUsersEnvelope": {
        "allOf": [
          {
            "$ref": "#/components/schemas/JsonSuccess"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "users"
                ],
                "properties": {
                  "users": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/AdminUser"
                    }
                  }
                },
                "additionalProperties": false
              }
            }
          }
        ]
      },
...
    }
  }
}